Combine msg server files for blind box mutation logic
Ref oursky/likecoin-chain#244
